Load fake chunks in order by distance
Previously fake chunks were loaded in the order in which `VisibleChunksTracker`
returned them (in rows from negative to positive X). Which in practice means
about half of all fake chunks need to be loaded before we get any chunk adjacent
to the real chunks (and can start rendering them). Not ideal.

This commit fixes that by sorting new loading jobs by distance to the player
chunk before submitting them to the executor.
It does not re-sort already submitted jobs because that should generally not be
of much use: If you move around, any newly loaded chunks will be at max distance
anyway, so adding them to the end of the queue is fine. And if you get
teleported around, existing jobs outside the view distance will be canceled
anyway.

Additionally it changes the `loadingJobs` hash map to be linked as to provide
the same order when we try to add loaded chunks into the game. Otherwise, if
the main thread work required to add them is resource starved, we'd load them in
hash map iteration order (which, again, would not be ideal).